extract common functions and consts
