+++ src/lib/consts.ts
@@ -1,0 +1,4 @@
+export const DEFAULT_NETWORK_URL = "ws://127.0.0.1:9944";
+export const DEFAULT_ASTAR_NETWORK_URL = "wss://rpc.astar.network";
+export const DEFAULT_SHIDEN_NETWORK_URL = "wss://rpc.shiden.astar.network";
+export const DEFAULT_SHIBUYA_NETWORK_URL = "wss://rpc.shibuya.astar.network";
